chore: update environment variables in `.env.example` and `oauth.py`
