+++ .env.example
@@ -2,3 +2,5 @@ SLACK_CLIENT_ID=''
 SLACK_CLIENT_SECRET=''
 DEV=true
 FLASK_SECRET_KEY=''
+SLACK_SIGNING_SECRET=''
+DEV_SLACK_REDIRECT_URI=''
